Organize tests, update settings, backend deps
