+++ .env
@@ -9,3 +9,4 @@ GITHUB_CLIENT_SECRET=
 FRONTEND_USE_IMAGE_PROXY=true
 FRONTEND_PWA=off
 SESSION_COOKIE_AGE=7776000
+PYTHONPATH=backend
