+++ .vscode/settings.json
@@ -34,14 +34,9 @@
         "editor.formatOnSave": true,
         "editor.defaultFormatter": "ms-python.black-formatter",
     },
-    "mypy.configFile": "backend/mypy.ini",
-    "mypy.runUsingActiveInterpreter": true,
-    "mypy.targets": [
-        "backend/coreapp",
-        "backend/decompme"
-    ],
     "typescript.tsdk": "frontend/node_modules/typescript/lib",
     "typescript.enablePromptUseWorkspaceTsdk": true,
-    "python.linting.mypyEnabled": true,
-    "python.linting.enabled": true
+    "mypy-type-checker.args": [
+        "--config-file=backend/mypy.ini"
+    ]
 }
