+++ backend/coreapp/views/project.py
@@ -6,24 +6,23 @@ from threading import Thread
 from typing import Any, Optional
 
 import django_filters
-from django.db.models.query import QuerySet
-from django.views import View
 from django.contrib.auth.models import User
+from django.db.models.query import QuerySet
 from django.db.utils import IntegrityError
-
+from django.views import View
 from github import Github, UnknownObjectException
-from github.Repository import Repository
 from github.GithubException import GithubException
+from github.Repository import Repository
 from rest_framework import filters, mixins, permissions, status
 from rest_framework.decorators import action
 from rest_framework.exceptions import APIException
 from rest_framework.pagination import CursorPagination
+from rest_framework.parsers import JSONParser, MultiPartParser
+from rest_framework.request import Request
 from rest_framework.response import Response
+from rest_framework.serializers import BaseSerializer
 from rest_framework.viewsets import GenericViewSet
 from rest_framework_extensions.routers import ExtendedSimpleRouter
-from rest_framework.request import Request
-from rest_framework.parsers import JSONParser, MultiPartParser
-from rest_framework.serializers import BaseSerializer
 
 from ..models.github import (
     GitHubRepo,
@@ -33,7 +32,6 @@ from ..models.github import (
 )
 from ..models.project import Project, ProjectFunction, ProjectMember
 from ..models.scratch import Scratch
-from ..models.profile import Profile
 from ..serializers import (
     ProjectFunctionSerializer,
     ProjectMemberSerializer,
